pix2pix/wandb/run-20241123_185748-81xuxy8x/files/requirements.txt
wheel==0.45.0
setuptools==75.3.0
pip==24.3.1
packaging==24.1
colorama==0.4.6
certifi==2024.8.30
idna==3.10
charset-normalizer==3.4.0
PySocks==1.7.1
pycparser==2.22
hpack==4.0.0
hyperframe==6.0.1
tqdm==4.67.0
h2==4.1.0
cffi==1.17.1
Brotli==1.1.0
zstandard==0.23.0
urllib3==2.2.3
requests==2.32.3
kagglehub==0.3.4
six==1.16.0
webencodings==0.5.1
text-unidecode==1.3
python-dateutil==2.9.0
bleach==6.2.0
python-slugify==8.0.4
kaggle==1.6.17
pillow==11.0.0
MarkupSafe==3.0.2
gmpy2==2.1.5
protobuf==5.27.5
filelock==3.16.1
networkx==3.2.1
typing_extensions==4.12.2
fsspec==2024.10.0
mpmath==1.3.0
Jinja2==3.1.4
sympy==1.13.3
numpy==1.26.4
wrapt==1.16.0
ml-dtypes==0.4.0
grpcio==1.65.5
tensorboard_data_server==0.7.0
torch==2.4.1.post102
absl-py==2.1.0
Werkzeug==3.1.3
zipp==3.21.0
importlib_metadata==8.5.0
Markdown==3.6
torchvision==0.19.1a0+5ed21bd
flatbuffers==24.3.25
opt_einsum==3.4.0
termcolor==2.5.0
astunparse==1.6.3
gast==0.5.5
google-pasta==0.2.0
cached-property==1.5.2
namex==0.0.8
Pygments==2.18.0
mdurl==0.1.2
tensorboard==2.17.1
markdown-it-py==3.0.0
rich==13.9.4
optree==0.13.1
h5py==3.12.1
keras==3.6.0
tensorflow==2.17.0
tensorflow_estimator==2.15.0
docker-pycreds==0.4.0
click==8.1.7
platformdirs==4.3.6
sentry-sdk==2.19.0
appdirs==1.4.4
smmap==5.0.0
gitdb==4.0.11
GitPython==3.1.43
psutil==6.1.0
setproctitle==1.3.4
PyYAML==6.0.2
wandb==0.18.7
